[change] Jenkinsfile내용 수정 - 자동 재시작
